requirements.txt
yfinance==0.2.27
pandas
numpy
matplotlib
streamlit
fastapi
uvicorn
langchain
openai
pymongo
alpaca-trade-api
stable-baselines3
torch
python-telegram-bot==20.0
pycoingecko==3.1.0
schedule==1.1.0
requests==2.31.0
plotly
feedparser
pandas-ta
